Refactored how paths/settings are determined for the different testing frameworks.
Instead of using a bunch case statements, there is now a module for each
framework that gets extended when you instantiate the Generator.

This makes it easier to add a new generator, since it'd be a matter of
just adding one module, rather than updating several places in
Generator.

Maybe this would lay the groundwork for custom generators?